requirements.txt
Flask==1.1.2
numpy==1.20.1
pandas==1.2.3
PyYAML==5.4.1
mysql-connector==2.2.9
